simplifed shard tensor signature

+++ src/transformers/core_model_loading.py
@@ -829,7 +829,7 @@ def spawn_tp_materialize(
     return a Callable that will load the tensor synchronously when called."""
 
     def _job():
-        sharded_tensor, _ = sharding_method.shard_tensor(tensor, param_casting_dtype=dtype, tensor_idx=tensor_idx)
+        sharded_tensor, _ = sharding_method.shard_tensor(tensor, dtype=dtype, tensor_idx=tensor_idx)
         return sharded_tensor
 
     if thread_pool is not None:
